HS-562: Add query by tenantId to repos
