+++ inventory/src/main/java/org/opennms/horizon/inventory/grpc/GrpcConfig.java
@@ -28,8 +28,7 @@
 
 package org.opennms.horizon.inventory.grpc;
 
-import java.util.Collections;
-
+import lombok.RequiredArgsConstructor;
 import org.opennms.horizon.inventory.mapper.MonitoringLocationMapper;
 import org.opennms.horizon.inventory.repository.MonitoringLocationRepository;
 import org.opennms.horizon.inventory.service.MonitoringGrpcService;
@@ -37,7 +36,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import lombok.RequiredArgsConstructor;
+import java.util.Collections;
 
 @RequiredArgsConstructor
 @Configuration
