+++ inventory/src/main/java/org/opennms/horizon/inventory/model/MonitoringLocation.java
@@ -1,7 +1,6 @@
 package org.opennms.horizon.inventory.model;
 
 import lombok.RequiredArgsConstructor;
-
 import org.hibernate.annotations.Type;
 
 import javax.persistence.Column;
