+++ inventory/src/main/java/org/opennms/horizon/inventory/repository/MonitoringSystemRepository.java
@@ -4,6 +4,10 @@ import org.opennms.horizon.inventory.model.MonitoringSystem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.UUID;
+
 @Repository
 public interface MonitoringSystemRepository extends JpaRepository<MonitoringSystem, Long> {
+    List<MonitoringSystem> findByTenantId(UUID tenantId);
 }
