+++ inventory/src/main/java/org/opennms/horizon/inventory/repository/NodeRepository.java
@@ -4,6 +4,10 @@ import org.opennms.horizon.inventory.model.Node;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.UUID;
+
 @Repository
 public interface NodeRepository extends JpaRepository<Node, Long> {
+    List<Node> findByTenantId(UUID tenantId);
 }
